% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo.R
\name{get_logo}
\alias{get_logo}
\title{Get logo images for teams}
\usage{
get_logo(teams)
}
\arguments{
\item{teams}{}
}
\value{
the image files for the teams
}
\description{
Get logo images for teams
}
\examples{
toronto <- get_logo("TOR")
plot(toronto)

}
