% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo.R
\name{logoInfo}
\alias{logoInfo}
\title{Get logo information for a teams}
\usage{
logoInfo(teams)
}
\arguments{
\item{teams}{a character vector of one or more \code{TeamID}s}
}
\value{
A data.frame with columns \code{teamID}, \code{name}, \code{png}
}
\description{
Get logo information for a teams
}
\examples{
logoInfo(c("TOR", "TEX"))

}
