% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoInfo.R
\name{logoInfo}
\alias{logoInfo}
\title{Get logo information for a teams}
\usage{
logoInfo(teams)
}
\arguments{
\item{teams}{A character vector of one or more \code{TeamID}s}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A data.frame with columns `teamID`, `name`, `png` and `image`
}\if{html}{\out{</div>}}
}
\description{
Get logo information for a teams
}
\examples{
logoInfo(c("TOR", "TEX"))

# get the file path to the Toronto Blue Jays logo
toronto <- logoInfo("TOR")[, "image"]
toronto

}
